[Testing] Reduce error verbosity of cleanup (#1365)
## Purpose ##
* When tests fail, they often show the error that occurs during cleanup,
rather than the error that caused the tests to fail
```
ERROR tests/llmcompressor/transformers/obcq/test_consecutive_runs.py::TestConsecutiveRunsSmall_0_commit::test_consecutive_runs_small - FileNotFoundError: [Errno 2] No such file or directory: './oneshot_output'
```

## Changes ## 
* Check if a directory exists before attempting to delete it during
cleanup

+++ tests/llmcompressor/transformers/finetune/test_oneshot_and_finetune_with_tokenizer.py
@@ -1,3 +1,4 @@
+import os
 import shutil
 import unittest
 
@@ -78,4 +79,5 @@ class TestOneshotAndFinetuneWithTokenizer(unittest.TestCase):
         print(tokenizer.decode(output[0]))
 
     def tearDown(self):
-        shutil.rmtree(self.output)
+        if os.path.isdir(self.output):
+            shutil.rmtree(self.output)
